+++ SourceCode/ChaosRayTracing/ChaosRayTracing/ChaosRayTracing.vcxproj
@@ -144,6 +144,7 @@
     <ClInclude Include="AABB.hpp" />
     <ClInclude Include="BVH.hpp" />
     <ClInclude Include="Camera.hpp" />
+    <ClInclude Include="Image.hpp" />
     <ClInclude Include="Light.hpp" />
     <ClInclude Include="Material.hpp" />
     <ClInclude Include="Math3D.hpp" />
